[fix] Ensure "global" is defined on the frontend
Currently, certain modules check global to make decisions about the
environment that its code is being executed in, but the "global"
variable is not present in the browser. One solution is to assign
the browser window object to "global". Not sure if this will
backfire, but seems to work for the module that complained.